[ODRP-344] Various changes requested by reviewer.
