+++ it/test/uk/gov/hmrc/dprs/BaseIntegrationSpec.scala
@@ -15,19 +15,24 @@
  */
 
 package uk.gov.hmrc.dprs
+import com.github.tomakehurst.wiremock.client.WireMock.getAllServeEvents
 import org.scalatest.concurrent.{IntegrationPatience, ScalaFutures}
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.matchers.{MatchResult, Matcher}
 import org.scalatestplus.play.guice.GuiceOneServerPerSuite
 import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.libs.ws.WSClient
+import play.api.libs.json.Json
+import play.api.libs.ws.{WSClient, WSResponse}
 import play.api.{Application, inject}
+import uk.gov.hmrc.dprs.BaseIntegrationSpec.CustomMatchers.{haveJsonBody, haveNoBody, haveStatus}
 import uk.gov.hmrc.dprs.services.AcknowledgementReferenceGenerator
 import uk.gov.hmrc.http.client.HttpClientV2
-import uk.gov.hmrc.http.test.HttpClientV2Support
+import uk.gov.hmrc.http.test.{HttpClientV2Support, WireMockSupport}
 
 import java.time.{Clock, Instant, ZoneId}
 import java.util.UUID
+import scala.jdk.CollectionConverters.CollectionHasAsScala
 
 class BaseIntegrationSpec
     extends AnyFreeSpec
@@ -35,23 +40,22 @@ class BaseIntegrationSpec
     with ScalaFutures
     with IntegrationPatience
     with GuiceOneServerPerSuite
-    with WireMockServerHandler
+    with WireMockSupport
     with HttpClientV2Support {
 
-  val wsClient: WSClient = app.injector.instanceOf[WSClient]
-  val baseUrl: String    = s"http://localhost:$port"
-  lazy val fixedClock: Clock  = Clock.fixed(Instant.now.truncatedTo(java.time.temporal.ChronoUnit.MILLIS), ZoneId.systemDefault)
+  val wsClient: WSClient                      = app.injector.instanceOf[WSClient]
+  val baseUrl: String                         = s"http://localhost:$port"
+  lazy val fixedClock: Clock                  = Clock.fixed(Instant.now.truncatedTo(java.time.temporal.ChronoUnit.MILLIS), ZoneId.systemDefault)
   lazy val fixedAcknowledgeReferenceGenerator = new FixedAcknowledgeReferenceGenerator(UUID.randomUUID().toString)
+  lazy val currentDateAndTime: String         = Instant.now(fixedClock).toString
 
-  override def fakeApplication(): Application = {
-    wireMockServer.start() // In order to get the dynamic port
+  override def fakeApplication(): Application =
     baseApplicationBuilder()
       .configure(extraApplicationConfig)
       .overrides(inject.bind[HttpClientV2].toInstance(httpClientV2))
       .overrides(inject.bind[Clock].toInstance(fixedClock))
       .overrides(inject.bind[AcknowledgementReferenceGenerator].toInstance(fixedAcknowledgeReferenceGenerator))
       .build()
-  }
 
   def extraApplicationConfig: Map[String, Any] = Map.empty
 
@@ -60,11 +64,66 @@ class BaseIntegrationSpec
       .configure(
         "metrics.enabled" -> false
       )
+
+  def fullUrl(path: String): String = baseUrl + "/dprs" + path
+
+  def verifyThatDownstreamApiWasCalled(): Unit =
+    getAllServeEvents.asScala.count(_.getWasMatched) shouldBe 1
+
+  def assertAsExpected(response: WSResponse, status: Int, jsonBodyOpt: Option[String] = None): Unit = {
+    response should haveStatus(status)
+    jsonBodyOpt match {
+      case Some(body) => response should haveJsonBody(body)
+      case None       => response should haveNoBody
+    }
+    ()
+  }
 }
 
 object BaseIntegrationSpec {
   class FixedAcknowledgeReferenceGenerator(generated: String) extends AcknowledgementReferenceGenerator {
     override def generate(): String = generated
   }
-}
 
+  object CustomMatchers {
+
+    class HaveStatus(expectedStatus: Int) extends Matcher[WSResponse] {
+      override def apply(response: WSResponse): MatchResult =
+        MatchResult(
+          response.status == expectedStatus,
+          s"We expected the response to have status [$expectedStatus], but it was actually [${response.status}].",
+          s"We didn't expect the response to have status [$expectedStatus], but it was indeed."
+        )
+    }
+
+    class HaveJsonBody(expectedRawJsonBody: String) extends Matcher[WSResponse] {
+      override def apply(response: WSResponse): MatchResult = {
+        val expectedJsonBody = Json.parse(expectedRawJsonBody)
+        val actualJsonBody   = Json.parse(response.body)
+        MatchResult(
+          actualJsonBody == expectedJsonBody,
+          s"We expected the response to have a json body [\n${Json.prettyPrint(expectedJsonBody)}\n], but it was actually [\n${Json.prettyPrint(actualJsonBody)}\n].",
+          s"We didn't expect the response to have a json body [\n${Json.prettyPrint(expectedJsonBody)}], but it was indeed."
+        )
+      }
+    }
+
+    class HaveNoBody() extends Matcher[WSResponse] {
+      override def apply(response: WSResponse): MatchResult = {
+        val actualBody = response.body
+        MatchResult(
+          actualBody == "",
+          s"We expected the response to have no body, but it was actually [\n$actualBody\n].",
+          s"We expected the response to have a body, but it didn't."
+        )
+      }
+    }
+
+    def haveStatus(expectedStatus: Int) = new HaveStatus(expectedStatus)
+
+    def haveJsonBody(expectedRawJsonBody: String) = new HaveJsonBody(expectedRawJsonBody)
+
+    def haveNoBody = new HaveNoBody()
+
+  }
+}
